Feature/abstraction

+++ src/main/java/educationManagementSystem/repository/UserRepository.java
@@ -2,12 +2,12 @@ package educationManagementSystem.repository;
 
 import java.util.Optional;
 
-import educationManagementSystem.model.User;
+import educationManagementSystem.model.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findById(Integer userId);
     Optional<User> findByUsername(String username);
